+++ src/events/eventHandler.ts
@@ -1,0 +1,10 @@
+import { CustomClient } from '../types/customClient';
+import { setupClientReadyHandler } from './clientReady';
+import { setupInteractionCreateHandler } from './interactionCreate';
+import { setupGuildMemberAddHandler } from './guildMemberAdd';
+
+export function setupEventHandlers(client: CustomClient) {
+    setupClientReadyHandler(client);
+    setupInteractionCreateHandler(client);
+    setupGuildMemberAddHandler(client);
+}
