+++ src/events/guildMemberAdd.ts
@@ -1,0 +1,9 @@
+import { Events } from 'discord.js';
+import { CustomClient } from '../types/customClient';
+
+export function setupGuildMemberAddHandler(client: CustomClient) {
+    client.on(Events.GuildMemberAdd, async member => {
+        member.send(`ようこそ ${member.displayName} さん！...`)
+            .catch(console.error);
+    });
+}
